backend/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
redis==7.0.1
aio-pika==9.5.7
openai==2.7.1
python-socketio==5.14.3
pydantic-settings==2.11.0
alembic==1.14.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pandas>=2.0.0
openpyxl==3.1.5
reportlab==4.2.5
boto3==1.35.99
firebase-admin==6.6.0
twilio==9.4.4
sendgrid==6.11.0
celery==5.4.0
flower==2.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2